+++ src/ScottPlot5/ScottPlot5/Plottables/AxisLine.cs
@@ -19,6 +19,7 @@ public abstract class AxisLine : IPlottable, IRenderLast
     public LineStyle LineStyle { get; set; } = new();
     public float LineWidth { get => LineStyle.Width; set => LineStyle.Width = value; }
     public LinePattern LinePattern { get => LineStyle.Pattern; set => LineStyle.Pattern = value; }
+    public Color LineColor { get => LineStyle.Color; set => LineStyle.Color = value; }
 
     public bool LabelOppositeAxis { get; set; } = false;