+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Rendering: Added support for gradient fills (#3298, #3157, #3310) _Thanks @KroMignon, @hockerschwan, and @faguetan_
 * Controls: Disabling interactions then re-enabling them restores original interactions (#3305, #3304) _Thanks @Nils-Berghs_
 * Plot: Added `Plot.GetPixel()` overload for improved support on multi-axis plots (#3306) _Thanks @MCF_
+* Axis Lines: Label background color may be distinct from line color (#3309) _Thanks @PhoenixChenLu_
 
 ## ScottPlot 5.0.20
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-21_